+++ magicctapipe/reco/estimators.py
